backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
fonttools==4.54.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
motor==3.7.0
numpy==2.1.2
optional-django==0.3.0
ordered-set==4.1.0
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.1
pillow==11.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.0
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
react==4.3.0
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
selenium==4.26.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.45.3
tailwindcss==0.0.1
trio==0.27.0
trio-websocket==0.11.1
typer==0.15.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.34.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0
wsproto==1.2.0
yara==1.7.7
yara-python==4.5.1
